+++ crates/hdk/CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
-## 0.4.0
+- Update `holochain_wasmer_guest`.
 
 ## 0.4.0-rc.1